requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
python-dotenv~=1.0.0
requests~=2.28.2
APScheduler~=3.10.1